requirements.txt
groq>=0.9.0
feedgen>=1.0.0
jinja2>=3.1.0
pyyaml>=6.0
